Fixed the layouts and organized the styles into a single folder.
