--- src/TopBar.js
@@ -1,9 +1,0 @@
-export default function TopBar() {
-  return (
-    <div className="top-bar">
-      <div className="logo-bar">
-        <h1 className="logo"></h1>
-      </div>
-    </div>
-  );
-}
