+++ src/index.js
@@ -3,10 +3,6 @@ import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom';
 import App from './App'
 
-
-
-
-
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
