+++ src/styles/due-styles.css
@@ -2,30 +2,32 @@
     box-sizing: border-box;
 }
 
-.Client-container > div.Client-header {
+.Due-container > div.Due-header {
     
     color: #222;
     padding: 10px 5px 0px;  
     font-size: 25px;
+   
     
 }
 
-.Client-content{
+.Due-content{
     display: grid;
     grid-template-columns: repeat(3, 1fr);
     gap: 5px;
     width: 1350px;
     height: 200px;
-    margin: auto;
+ 
+    padding: 0px 10px 10px;
 }
-.Agents{
+.Due-total{
     background: white;
     padding: 20px;
     background-color: lightgray;
     font-size: 20px;
 }
 
-.Agents > p {
+.Due-total > p {
     text-align: center;
     font-size: 20px;
 }
