+++ src/styles/nav-styles.css
@@ -48,7 +48,7 @@ body, html {
    background-color: #ffffff;
 }
 
-@media screen and (max-width: 1920px) {
+@media screen and (min-width: 1080px) and (max-width: 2000px) {
    .nav-bar {
       width: 180px; 
    }
@@ -69,7 +69,7 @@ body, html {
    }
 }
 
-@media screen and (max-height: 900px)  {
+@media screen and (min-height: 300px) and (max-height: 900px) {
    .nav-bar {
       width: 180px; 
    }
