--- src/top-styles.css
@@ -1,21 +1,0 @@
-*{
-    box-sizing: border-box;
-}
-
-.top-bar {
-    position: fixed;
-    background-color: #8A7F7F;
-    height: 80px;
-    width: 100%;  
- }
- 
- .top-bar .logo-bar
-{
-  
-   height: 100px;
-   width: 100%;
-   display: flex;
-   align-items: center;
-   justify-content: center;
-   margin-left: 20px;
-}
